+++ components/Calendar/DatePicker.tsx
@@ -1,12 +1,12 @@
 import { NextPage } from 'next'
-import { ChangeEventHandler, Dispatch, SetStateAction, useState } from 'react'
+import { Dispatch, SetStateAction } from 'react'
 import { CalChangeHandler } from '../../pages/cal'
 import {
   DateDisplayStyle,
   DateNumStyle,
   DateSepStyle,
   YearNumStyle,
-} from './DatePicker.style'
+} from './Calendar.style'
 
 interface Props {
   month: number
@@ -25,9 +25,15 @@ const DatePicker: NextPage<Props> = ({
 }) => {
   return (
     <DateDisplayStyle>
-      <YearNumStyle value={year} onChange={handleChange(setYear, 'year')} />
-      <DateSepStyle>.</DateSepStyle>
-      <DateNumStyle value={month} onChange={handleChange(setMonth, 'month')} />
+      <div>
+        <YearNumStyle value={year} onChange={handleChange(setYear, 'year')} />
+        <DateSepStyle>.</DateSepStyle>
+        <DateNumStyle
+          value={month}
+          onChange={handleChange(setMonth, 'month')}
+        />
+      </div>
+      <div>이동</div>
     </DateDisplayStyle>
   )
 }
